Quadplot means, original

Makes a quadplot but does NOT standardize data first. Requires that you use NormalizeTimeSeries separately.

QuadPlot_means(
  DATA = NULL,
  D1 = NA,
  D2 = NA,
  SE.d1.up = NULL,
  SE.d1.lo = NULL,
  SE.d2.up = NULL,
  SE.d2.lo = NULL,
  SE.d1 = NULL,
  SE.d2 = NULL,
  d1.col = NULL,
  d2.col = NULL,
  rng = NULL,
  Xlim = NA,
  Ylim = NA,
  Title = NA,
  plot.type = 1,
  style = 1,
  flip.colors = NULL,
  bg.polys = FALSE,
  Xlab = NULL,
  Ylab = NULL,
  Label = NA,
  Symbol = c(21, 21, 21, 21, 21, 24, 24, 24, 24, 24, 25, 25, 25, 25, 25),
  bg.color = c("black", "white", "grey", "red", "blue"),
  ol.color = "black",
  AxisCex = 0.9,
  LabelCex = 0.9,
  XYLabCex = 0.9,
  LegendCex = 0.8,
  LegendPointCex = 1,
  TitleCex = 1,
  TextCex = 0.8,
  PointCex = 1,
  error.lwd = NA,
  PERIOD = 5,
  Period.Label = "Years",
  Bmar = NA,
  Lmar = NA,
  Tmar = NA,
  Rmar = NA,
  AltTxt = NA,
  TextOffset = 0.5,
  PlotLegend = NULL
)

Arguments

DATA

...

D1

...

D2

...

SE.d1.up

...

SE.d1.lo

...

SE.d2.up

...

SE.d2.lo

...

SE.d1

...

SE.d2

...

d1.col

...

d2.col

...

rng

...

Xlim

...

Ylim

...

Title

...

plot.type

...

style

...

flip.colors

...

bg.polys

...

Xlab

...

Ylab

...

Label

...

Symbol

...

bg.color

...

ol.color

...

AxisCex

...

LabelCex

...

XYLabCex

...

LegendCex

...

LegendPointCex

...

TitleCex

...

TextCex

...

PointCex

...

error.lwd

...

PERIOD

...

Period.Label

...

Bmar

...

Lmar

...

Tmar

...

Rmar

...

AltTxt

...

TextOffset

...

PlotLegend

...

Value

Details

x = short term trend y = mean over the evaluation period the data come out of the normalization in this format style 1 will graph them as x,y style 2 will swap the axes internally to y, x. this is done to maintain some formatting between some different codes CAN plot standard errors but not clear what they should be at present. no alt text for standard errors. PlotLegend can be used to turn the Legend on or Off. (TRUE or FALSE)