CCIEA Anthropogenic Drivers

Commercial shipping - distance: source data: Domestic vessel data from U.S. Army Corps of Engineers Navigation Data Center (New Orleans, LA) and foreign vessel data from http://www.ndc.iwr.usace.army.mil/data/dataclen.htm ., additional calculations: Commercial shipping activity (distance) was measured as the distance traveled by commercial vessels during transit within waters of the California Current. Distance traveled was calculated using distance traveled within the California Current while in transit between shipping and receiving ports; Finfish Aquaculture: source data: Washington Department of Fish & Wildlife, Commercial Harvest Data Team, additional calculations: Finfish aquaculture was measured as the production of finfish from aquaculture operations that are located in marine waters. Only the State of Washington currently operates net-pen farms for Atlantic salmon (Salmo salar).; Nutrient Input: source data: We used county-level data from 1987 - 2006 and state-level data from 2007 - 2010 from the U.S. Geological Survey (Ruddy et al. 2006, Gronberg and Spahr 2012; http://water.usgs.gov/lookup/getspatial?sir2012-5207_county_fertilizer) and nationwide data (1945 - 2001; Ruddy et al. (2006)) to develop an index for the California Current across the longer time series., additional calculations: Nutrient input was measured using a normalized index of total nitrogen and phosphorus input from agricultural fertilizers used within watersheds that drain into the California Current.; Oil And Gas Activity: source data: Oil production data come from annual reports of the California State Department of Conservation's Division of oil, gas, and geothermal resources (ftp://ftp.consrv.ca.gov/../pub/oil/annual_reports/), while natural gas production come from the U.S. Energy Information Administration (http://www.eia.gov/dnav/ng/ng_prod_sum_dcu_rcatf_a.htm)., additional calculations: Oil and gas activities were measured using a normalized index combining the production of oil and natural gas occurring in offshore sites of California.; Seafood Demand: source data: Data can be found in NOAA's annual Fisheries of the United States reports (http://www.st.nmfs.noaa.gov/st1/publications.html)., additional calculations: Seafood demand was measured as total consumption or utilization of edible and non-edible fisheries products across the entire United States. We used data from the entire United States as seafood from the California Current is consumed and utilized at national, and even international scales.; Seafood Demand Per Capita: source data: Data can be found in NOAA's annual Fisheries of the United States reports (http://www.st.nmfs.noaa.gov/st1/publications.html)., additional calculations: Seafood demand was measured as total consumption or utilization of edible and non-edible fisheries products across the entire United States. We used data from the entire United States as seafood from the California Current is consumed and utilized at national, and even international scales.; Shellfish Aquaculture: source data: Data were retrieved and summed together from Washington Department of Fish and Wildlife's Commercial Harvest Data Team, Oregon Department of Agriculture and the California Department of Fish and Game., additional calculations: Shellfish aquaculture was measured as the estimated production of shellfish from aquaculture operations in the states of Washington, Oregon and California.; Bottom trawl contact with seafloor habitat (weighted): source data: Distances were summed within physiographic depth, habitat and ecoregion categories from logbook data provided by the West Coast Groundfish Observer Program at the Northwest Fisheries Science Center and is comparable to the data developed for NOAA's Essential Fish Habitat 5-year Synthesis Review in 2013 (maps and data available: http://efh-catalog.coas.oregonstate.edu/overview/, additional calculations: Habitat modification was measured using the total distance disturbed by trawling and fixed (longlines and pots) gear. Straight line distances between start and end points for trawling gear and between set and retrieval points for fixed gear were calculated for each gear type and weighted by the gear type's impact to the bottom habitat and by the type of habitat. These weightings come from NOAA's 5-year Synthesis Review of West Coast Groundfish (Supplemental Table A3a.2); Groundfish landings (w/o hake) WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) landings (1000's of metric tons) in Washington. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Coastal pelagic species (w/o squid) landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) landings (1000's of metric tons) on the U.S. West Coast. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Coastal pelagic species (w/o squid) landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Coastal pelagic species (w/o squid) landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) landings (1000's of metric tons) in Oregon. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Coastal pelagic species (w/o squid) landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Crab landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab landings (1000's of metric tons) on the U.S. West Coast. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Crab landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab landings (1000's of metric tons) in California. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Crab landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab landings (1000's of metric tons) in Oregon. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Crab landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab landings (1000's of metric tons) in Washington. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Groundfish landings (w/o hake) coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) landings (1000's of metric tons) on the U.S. West Coast. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Groundfish landings (w/o hake) CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) landings (1000's of metric tons) in California. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Groundfish landings (w/o hake) OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) landings (1000's of metric tons) in Oregon. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Highly migratory species landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species landings (1000's of metric tons) on the U.S. West Coast. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Highly migratory species landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species landings (1000's of metric tons) in California. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Highly migratory species landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species landings (1000's of metric tons) in Oregon. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Highly migratory species landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species landings (1000's of metric tons) in Washington. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Market squid landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Market squid landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Market squid (Loligo opalescens) landings (1000's of metric tons) in California.; Market squid landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Other species landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species landings (1000's of metric tons) on the U.S. West Coast. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Other species landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species landings (1000's of metric tons) in California. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Other species landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species landings (1000's of metric tons) in Oregon. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Other species landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species landings (1000's of metric tons) in Washington. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Total Fisheries landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries landings (1000's of metric tons) in Washington.; Pacific hake landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) landings (1000's of metric tons) on the U.S. West Coast. Pacific hake landings include data from shoreside and at-sea processors.; Pacific hake landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) landings (1000's of metric tons) in California. Pacific hake landings include data from shoreside and at-sea processors.; Pacific hake landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) landings (1000's of metric tons) in Oregon. Pacific hake landings include data from shoreside and at-sea processors.; Pacific hake landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) landings (1000's of metric tons) in Washington. Pacific hake landings include data from shoreside and at-sea processors.; Recreational landings coastwide: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Total landings of all species from recreational fisheries from www.recfin.org using weight of catch type "A + B1" metric tons.; Recreational landings CA: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Total landings of all species from recreational fisheries in California from www.recfin.org using weight of catch type "A + B1" metric tons.; Recreational landings OR: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Total landings of all species from recreational fisheries in Oregon from www.recfin.org using weight of catch type "A + B1" metric tons.; Recreational landings WA: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Total landings of all species from recreational fisheries in Washington from www.recfin.org using weight of catch type "A + B1" metric tons.; Salmon commercial landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon landings (1000's of metric tons) on the U.S. West Coast. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon commercial landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon landings (1000's of metric tons) in California. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon commercial landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon landings (1000's of metric tons) in Oregon. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon commercial landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon landings (1000's of metric tons) in Washington. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon recreational landings coastwide: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Salmon landings (1000's of metric tons) on the U.S. West Coast. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon recreational landings CA: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Salmon landings (1000's of metric tons) in California. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon recreational landings OR: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Salmon landings (1000's of metric tons) in Oregon. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Salmon recreational landings WA: source data: Recreational Fisheries Information Network (RecFIN; http://www.recfin.org/), additional calculations: Salmon landings (1000's of metric tons) in Washington. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Shrimp landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp landings (1000's of metric tons) on the U.S. West Coast. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Shrimp landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp landings (1000's of metric tons) in California. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Shrimp landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp landings (1000's of metric tons) in Oregon. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Shrimp landings WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp landings (1000's of metric tons) in Washington. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Total Fisheries landings coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries landings (1000's of metric tons) on the U.S. West Coast.; Total Fisheries landings CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries landings (1000's of metric tons) in California.; Total Fisheries landings OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries landings (1000's of metric tons) in Oregon.; Groundfish revenue (w/o hake) coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) on the U.S. West Coast. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Coastal pelagic species (w/o squid) revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) revenue (millions of 2015 dollars) on the U.S. West Coast. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Shrimp revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp revenue (millions of 2015 dollars) on the U.S. West Coast. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Salmon revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon revenue (millions of 2015 dollars) on the U.S. West Coast. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Other species revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species revenue (millions of 2015 dollars) on the U.S. West Coast. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Pacific hake revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) on the U.S. West Coast. Pacific hake landings include data from shoreside and at-sea processors.; Market squid revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Crab revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab revenue (millions of 2015 dollars) on the U.S. West Coast. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Highly migratory species revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species revenue (millions of 2015 dollars) on the U.S. West Coast. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Commercial fisheries revenue coastwide: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries revenue (millions of 2015 dollars) on the U.S. West Coast.; Groundfish revenue (w/o hake) CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in California. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Coastal pelagic species (w/o squid) revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) revenue (millions of 2015 dollars) in California. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Shrimp revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp revenue (millions of 2015 dollars) in California. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Salmon revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon revenue (millions of 2015 dollars) in California. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Other species revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species revenue (millions of 2015 dollars) in California. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Pacific hake revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in California. Pacific hake landings include data from shoreside and at-sea processors.; Market squid revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Crab revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab revenue (millions of 2015 dollars) in California. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Highly migratory species revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species revenue (millions of 2015 dollars) in California. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Commercial fisheries revenue CA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries revenue (millions of 2015 dollars) in California.; Groundfish revenue (w/o hake) OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in Oregon. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Coastal pelagic species (w/o squid) revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) revenue (millions of 2015 dollars) in Oregon. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Shrimp revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp revenue (millions of 2015 dollars) in Oregon. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Salmon revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon revenue (millions of 2015 dollars) in Oregon. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Other species revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species revenue (millions of 2015 dollars) in Oregon. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Pacific hake revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in Oregon. Pacific hake landings include data from shoreside and at-sea processors.; Market squid revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/); Crab revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab revenue (millions of 2015 dollars) in Oregon. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Highly migratory species revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species revenue (millions of 2015 dollars) in Oregon. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Commercial fisheries revenue OR: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries revenue (millions of 2015 dollars) in Oregon.; Groundfish revenue (w/o hake) WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Groundfish (without Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in Washington. Groundfish taxa include flatfishes, rockfishes and abundant demersal roundfishes.; Coastal pelagic species (w/o squid) revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Coastal pelagic species (without market squid (Loligo opalescens)) revenue (millions of 2015 dollars) in Washington. Coastal pelagic species include Pacific herring (Clupea harengus pallasii), round herring (Etrumeus teres), chub mackerel (Scomber japonicus), jack mackerel (Trachurus symmetricus), northern anchovy (Engraulis mordax), Pacific bonito (Sarda chiliensis), Pacific sardine (Sardinops sagax), and unspecified mackerel.; Shrimp revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Shrimp revenue (millions of 2015 dollars) in Washington. Shrimp landings consist primarily of Pacific pink shrimp (Pandalus jordani).; Salmon revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Salmon revenue (millions of 2015 dollars) in Washington. Salmon landings primarily consist of Chinook (Oncorhynchus tshawytscha), but also includes chum (O. keta), coho (O. kisutch), pink (O. gorbuscha) and sockeye (O. nerka) and steelhead (O. mykiss) .; Other species revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Other species revenue (millions of 2015 dollars) in Washington. Other species include several taxa, but consists primarily of red sea urchin (Stronglyocentrotus franciscanus) and hagfish (Eptatretus spp.).; Pacific hake revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Pacific hake (Merluccius productus)) revenue (millions of 2015 dollars) in Washington. Pacific hake landings include data from shoreside and at-sea processors.; Crab revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Crab revenue (millions of 2015 dollars) in Washington. Crab species include Dungeness (Metacarcinus magister), tanner (Chionoecetes spp.), rock (Cancer spp.) and unspecified crabs.; Highly migratory species revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Highly migratory species revenue (millions of 2015 dollars) in Washington. Highly migratory species primarily consist of tunas (Thunnus spp), swordfish (Xiphias gladius) and pelagic sharks (e.g., blue (Prionace glauca), thresher (Alopias spp), and shortfin mako (Isurus oxyrinchus).; Commercial Fisheries revenue WA: source data: Pacific Fisheries Information Network (PacFIN; http://pacfin.psmfc.org/), additional calculations: Combined commercial and recreational fisheries revenue (millions of 2015 dollars) in Washington.;

cciea_AC

Format

A data frame with 75 rows and 95 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) -7.889184E8, 1.5463008E9

shipping_distance_traveled

Commercial shipping - distance (Millions of km) 17.68676, 31.40238

finfish_production

Finfish Aquaculture (1000s mt) 0.01149992, 10.43139

fertilizer_applications

Nutrient Input (Normalized index) -1.867494, 1.818944

oil_gas_production

Oil And Gas Activity (Normalized index) -2.330674, 1.640257

consumption_fish

Seafood consumption (total) (Millions of metric tons) 4.482482, 10.0965

consumption_per_cap

Seafood consumption (per capita) (kg) 20.95595, 39.28107

shellfish_production

Shellfish Aquaculture (metric tons) 5.537665, 11.91462

habitat_modification

Bottom trawl contact with seafloor habitat (weighted) (Weighted 1000s km) 76.4527, 373.32

groundfish_landings_wa

Groundfish landings (w/o hake) WA (1000s metric tons) 3.268318, 30.97578

cps_landings_coastwide

Coastal pelagic species (w/o squid) landings coastwide (1000s metric tons) 10.70933, 146.6982

cps_landings_ca

Coastal pelagic species (w/o squid) landings CA (1000s metric tons) 9.179026, 116.2179

cps_landings_or

Coastal pelagic species (w/o squid) landings OR (1000s metric tons) 0.1021285, 45.57781

cps_landings_wa

Coastal pelagic species (w/o squid) landings WA (1000s metric tons) 0.4017799, 36.98861

crab_landings_coastwide

Crab landings coastwide (1000s metric tons) 4.784686, 41.72063

crab_landings_ca

Crab landings CA (1000s metric tons) 1.777526, 15.26942

crab_landings_or

Crab landings OR (1000s metric tons) 1.720652, 15.29602

crab_landings_wa

Crab landings WA (1000s metric tons) 1.286507, 14.47761

groundfish_landings_coastwide

Groundfish landings (w/o hake) coastwide (1000s metric tons) 23.25325, 124.748

groundfish_landings_ca

Groundfish landings (w/o hake) CA (1000s metric tons) 5.847436, 52.65097

groundfish_landings_or

Groundfish landings (w/o hake) OR (1000s metric tons) 9.699529, 41.12125

hms_landings_coastwide

Highly migratory species landings coastwide (1000s metric tons) 7.694124, 54.95815

hms_landings_ca

Highly migratory species landings CA (1000s metric tons) 0.6113654, 52.93386

hms_landings_or

Highly migratory species landings OR (1000s metric tons) 0.4897378, 4.879423

hms_landings_wa

Highly migratory species landings WA (1000s metric tons) 0.06658192, 10.78039

squid_landings_coastwide

Market squid landings coastwide (1000s metric tons) 0.9934503, 130.8523

squid_landings_ca

Market squid landings CA (1000s metric tons) 0.5640231, 130.8447

squid_landings_or

Market squid landings OR (1000s metric tons) 0.0, 3.196166

other_species_landings_coastwide

Other species landings coastwide (1000s metric tons) 6.098473, 37.44385

other_species_landings_ca

Other species landings CA (1000s metric tons) 2.922818, 26.98719

other_species_landings_or

Other species landings OR (1000s metric tons) 0.4558114, 4.70202

other_species_landings_wa

Other species landings WA (1000s metric tons) 1.004735, 7.22235

total_fisheries_landings_wa

Total Fisheries landings WA (1000s metric tons) 72.43497, 177.1616

pacific_hake_landings_coastwide

Pacific hake landings coastwide (1000s metric tons) 69.49753, 354.2313

pacific_hake_landings_ca

Pacific hake landings CA (1000s metric tons) 0.003771167, 135.1784

pacific_hake_landings_or

Pacific hake landings OR (1000s metric tons) 14.8115, 241.9501

pacific_hake_landings_wa

Pacific hake landings WA (1000s metric tons) 6.960765, 142.5702

recreational_landings_coastwide

Recreational landings coastwide (1000s metric tons) 5.09321, 8.838449

recreational_landings_ca

Recreational landings CA (1000s metric tons) 3.214205, 6.593854

recreational_landings_or

Recreational landings OR (1000s metric tons) 0.555653, 1.268657

recreational_landings_wa

Recreational landings WA (1000s metric tons) 0.4506299, 1.24569

salmon_com_landings_coastwide

Salmon commercial landings coastwide (1000s metric tons) 5.131682, 37.98297

salmon_com_landings_ca

Salmon commercial landings CA (1000s metric tons) 6.09329E-4, 7.70475

salmon_com_landings_or

Salmon commercial landings OR (1000s metric tons) 0.4526994, 8.078773

salmon_com_landings_wa

Salmon commercial landings WA (1000s metric tons) 3.318938, 29.93654

salmon_rec_landings_coastwide

Salmon recreational landings coastwide (1000s fish) 47.0, 774.0

salmon_rec_landings_ca

Salmon recreational landings CA (1000s fish) 0.0, 398.0

salmon_rec_landings_or

Salmon recreational landings OR (1000s fish) 6.0, 305.0

salmon_rec_landings_wa

Salmon recreational landings WA (1000s fish) 23.0, 323.0

shrimp_landings_coastwide

Shrimp landings coastwide (1000s metric tons) 5.660932, 47.73951

shrimp_landings_ca

Shrimp landings CA (1000s metric tons) 0.5452471, 9.112365

shrimp_landings_or

Shrimp landings OR (1000s metric tons) 2.249487, 24.30083

shrimp_landings_wa

Shrimp landings WA (1000s metric tons) 1.424119, 19.11796

total_fisheries_landings_coastwide

Total Fisheries landings coastwide (1000s metric tons) 337.4346, 569.5687

total_fisheries_landings_ca

Total Fisheries landings CA (1000s metric tons) 51.71044, 302.277

total_fisheries_landings_or

Total Fisheries landings OR (1000s metric tons) 68.72131, 290.9759

groundfish_revenue_coastwide

Groundfish revenue (w/o hake) coastwide (Millions $ (year 2015)) 52.97257, 163.5798

cps_revenue_coastwide

Coastal pelagic species (w/o squid) revenue coastwide (Millions $ (year 2015)) 2.481962, 67.00124

shrimp_revenue_coastwide

Shrimp revenue coastwide (Millions $ (year 2015)) 15.77684, 109.9221

salmon_revenue_coastwide

Salmon revenue coastwide (Millions $ (year 2015)) 22.38714, 335.2793

other_species_revenue_coastwide

Other species revenue coastwide (Millions $ (year 2015)) 28.03679, 105.3409

pacific_hake_revenue_coastwide

Pacific hake revenue coastwide (Millions $ (year 2015)) 0.9633493, 74.55263

squid_revenue_coastwide

Market squid revenue coastwide (Millions $ (year 2015)) 1.237159, 83.42967

crab_revenue_coastwide

Crab revenue coastwide (Millions $ (year 2015)) 28.39146, 247.6923

hms_revenue_coastwide

Highly migratory species revenue coastwide (Millions $ (year 2015)) 24.45324, 151.4093

total_fisheries_revenue_coastwide

Commercial fisheries revenue coastwide (Millions $ (year 2015)) 361.1931, 858.0573

groundfish_revenue_ca

Groundfish revenue (w/o hake) CA (Millions $ (year 2015)) 17.22728, 73.89497

cps_revenue_ca

Coastal pelagic species (w/o squid) revenue CA (Millions $ (year 2015)) 2.157374, 64.81139

shrimp_revenue_ca

Shrimp revenue CA (Millions $ (year 2015)) 4.883509, 19.83477

salmon_revenue_ca

Salmon revenue CA (Millions $ (year 2015)) 0.0, 90.68086

other_species_revenue_ca

Other species revenue CA (Millions $ (year 2015)) 22.35475, 81.14681

pacific_hake_revenue_ca

Pacific hake revenue CA (Millions $ (year 2015)) 1.7603E-4, 2.298333

squid_revenue_ca

Market squid revenue CA (Millions $ (year 2015)) 0.7450277, 83.42967

crab_revenue_ca

Crab revenue CA (Millions $ (year 2015)) 10.57268, 109.2162

hms_revenue_ca

Highly migratory species revenue CA (Millions $ (year 2015)) 2.774751, 145.0685

total_fisheries_revenue_ca

Commercial fisheries revenue CA (Millions $ (year 2015)) 134.3722, 363.4424

groundfish_revenue_or

Groundfish revenue (w/o hake) OR (Millions $ (year 2015)) 19.44821, 54.411

cps_revenue_or

Coastal pelagic species (w/o squid) revenue OR (Millions $ (year 2015)) 0.006917352, 10.19605

shrimp_revenue_or

Shrimp revenue OR (Millions $ (year 2015)) 5.57979, 68.78595

salmon_revenue_or

Salmon revenue OR (Millions $ (year 2015)) 2.516938, 84.31818

other_species_revenue_or

Other species revenue OR (Millions $ (year 2015)) 1.256084, 10.05517

pacific_hake_revenue_or

Pacific hake revenue OR (Millions $ (year 2015)) 1.79066E-4, 58.99173

squid_revenue_or

Market squid revenue OR (Millions $ (year 2015)) 0.0, 3.12914

crab_revenue_or

Crab revenue OR (Millions $ (year 2015)) 10.856, 75.87123

hms_revenue_or

Highly migratory species revenue OR (Millions $ (year 2015)) 1.828291, 21.32934

total_fisheries_revenue_or

Commercial fisheries revenue OR (Millions $ (year 2015)) 82.86972, 211.2898

groundfish_revenue_wa

Groundfish revenue (w/o hake) WA (Millions $ (year 2015)) 8.452095, 41.17684

cps_revenue_wa

Coastal pelagic species (w/o squid) revenue WA (Millions $ (year 2015)) 0.2302786, 9.370674

shrimp_revenue_wa

Shrimp revenue WA (Millions $ (year 2015)) 4.136456, 35.50843

salmon_revenue_wa

Salmon revenue WA (Millions $ (year 2015)) 7.821066, 182.577

other_species_revenue_wa

Other species revenue WA (Millions $ (year 2015)) 2.017696, 17.0246

pacific_hake_revenue_wa

Pacific hake revenue WA (Millions $ (year 2015)) 0.02971284, 38.29011

crab_revenue_wa

Crab revenue WA (Millions $ (year 2015)) 6.962783, 104.1098

hms_revenue_wa

Highly migratory species revenue WA (Millions $ (year 2015)) 0.2235326, 31.69628

total_fisheries_revenue_wa

Commercial Fisheries revenue WA (Millions $ (year 2015)) 127.8284, 306.4503

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_AC/index.html