Seabird beached mortality

Annual mean encounter rates (bird carcasses/km surveyed) aggregated from October to February (inclusive, with years labelled according to the convention that Oct 2014 to Feb 2015 are labelled as 2014) for each surveyed beach location, and then averaged across all beaches surveyed in that year. No CC: Data from the Coastal Observation and Seabird Survey Team (COASST), who coordinate a team of trained volunteers that collect effort-controlled survey data on an approximately monthly basis, recording beached bird numbers and identity from survey locations in Northern California through to Northern Washington and into Alaska and the Bering Sea. Contact COASST (https://depts.washington.edu/coasst/) for details on calculations before citing or distributing these data. So/Ce CC: Data from BeachCombers, who coordinate a team of trained volunteers that collect effort-controlled survey data on an approximately monthly basis, recording beached bird numbers and identity from survey locations in Central/Southern California. Contact BeachCombers (https://www.mlml.calstate.edu/beachcombers/) for details on calculations before citing or distributing these data. Ce CC: Data from BeachWatch, who coordinate a team of trained volunteers that collect effort-controlled survey data on an approximately monthly basis, recording beached bird numbers and identity from survey locations in Central California. Contact BeachWatch (https://farallones.noaa.gov/science/beachwatch.html) for details on calculations before citing or distributing these data.

cciea_B_B_MORT

Format

A data frame with 527 rows and 3 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) [7.573824E8, 1.5463008E9]

encounter_rate

Encounter Rate () [0.0, 31.652]

species_cohort

Species (region, season) () []

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_B_B_MORT/index.html