Domoic acid levels in razor clams and other species, West Coast

Data on domoic acid concentrations in razor clams are from Ms. Audrey Coyne (audrey.coyne@doh.wa.gov; Washington State Department of Health); these data are compiled from tests conducted by a variety of Tribal, State, and County partners on Washington beaches. Domoic acid levels are reported in ppm. Sample testing frequency is irregular as it depends on the timing of proposed recreational razor clamming digs by Washington State Department of Fish and Wildlife and prevalence of recent detections. We present data as maximum monthly values compiled from samples in a variety of locations on the Washington State outer coast beaches.

cciea_EI_DOMACID

Format

A data frame with 19178 rows and 6 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) -3.643488E8, 1.5740352E9

site

Site Name () []

domoic_acid_concentration

Domoic Acid Concentration (ppm) 0.0, 730.0

water_body

Water Body () []

region

Region () []

species

Species () []

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_DOMACID/index.html