Groundfish mean trophic level

Source Data: Groundfish data are from the NMFS U.S. West Coast Groundfish Bottom Trawl Survey (https://www.nwfsc.noaa.gov/research/divisions/fram/groundfish/bottom_trawl.cfm) and were downloaded from the FRAM data warehouse at https://www.nwfsc.noaa.gov/data/map with the following API query:Â https://www.nwfsc.noaa.gov/data/api/v1/source/trawl.catch_fact/selection.csv?variables=program,trawl_id,date_dim$year,date_dim$yyyymmdd,vessel,performance,year_stn_invalid,depth_m,latitude_dd,longitude_dd,scientific_name,common_name,species_category,partition,total_catch_numbers,total_catch_wt_kg,cpue_kg_per_ha_der,cpue_numbers_per_ha_der,"actual_station_design_dim$station_invalid_survey_year", "actual_station_design_dim$reason_station_invalid","reason_stn_invalid (target station)", "year_stn_invalid" Additional Calculations: Area-weighted mean trophic level (MTL) for West Coast groundfishes. MTL was calculated for survey data without adjusting for sampling effort in different depth x latitude strata. For details, see Tolimieri et al. 2014 (https://www.noaa.gov/iea/CCIEA-Report/pdf/index.html)

cciea_EI_MTL

Format

A data frame with 48 rows and 5 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) 1.0413792E9, 1.5147648E9

population

Population () []

mean_trophic_level

() 3.614351, 3.773371

Seup

Confidence Interval, Upper () 3.623129, 3.786633

Selo

Confidence Interval, Lower () 3.605017, 3.760109

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_MTL/index.html