Rockfish Recruitment and Ecosystem Assessement Survey, Biodiversity: Grid format

Biodiversity indices (total abundance, species richness, Shannon-Weaver diversity, Pielou's evenness) of all taxa (also young-of-the-year rockfish and groundfish taxa, and forage taxa) caught in the RREAS (Rockfish Recruitment and Ecosystem Assessment Survey) mid-water trawl. This is a regional average of all net haul stations located from Point Reyes to Monterey CA.

cciea_EI_RREAS_diversity_grid

Format

A data frame with 87 rows and 10 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) 6.31152E8, 1.5147648E9

taxa

Taxa () []

abundance

Total Abundance (ln(CPUE + 1)) 1.390595, 12.2398

abundance_error

Abundance Error (ln(CPUE + 1)) 1.744331, 13.34593

richness

Species Richness (number of species) 1.5, 21.98214

richness_error

Species Richness Error (number of species) 1.453661, 7.003501

diversity

Shannon-Weaver Diversity () 0.0696443, 1.690858

diversity_error

Shannon-Weaver Diversity Error () 0.1325587, 0.8078451

evenness

Pielou's Evenness () 0.03166048, 0.7764255

evenness_error

Pielou's Evenness Error () 0.05746754, 0.4711267

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_RREAS_diversity_grid/index.html