Rockfish Recruitment and Ecosystem Assessement Survey, Biodiversity: CCIEA format

Biodiversity indices (total abundance, species richness, Shannon-Weaver diversity, Pielou's evenness) of all taxa (also young-of-the-year rockfish and groundfish taxa, and forage taxa) caught in the RREAS (Rockfish Recruitment and Ecosystem Assessment Survey) mid-water trawl. This is a regional average of all net haul stations located from Point Reyes to Monterey CA.

cciea_EI_RREAS_diversity_list

Format

A data frame with 348 rows and 7 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) 6.31152E8, 1.5147648E9

metric

() []

timeseries

() []

data

() 0.03166048, 21.98214

error

() 0.05746754, 13.34593

SEup

() 0.08912803, 27.93191

SElo

() -1.839871, 16.03238

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_RREAS_diversity_list/index.html