Highly Migratory Species

This index represents modeled female spawning potential biomass from the latest (2017) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). North Pacific albacore are considered to be one stock throughout the North Pacific Ocean, although some studies have suggested that a northern and southern stock may be present within the assessment area. They are fished throughout their range by multiple countries, mostly using surface gear (troll and pole and line), as well as pelagic longlines and other gears. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Estimates of recruitment are derived from the assessment model. The latest stock assessment report was completed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. The assessment assumes that there is one stock of bigeye in the eastern Pacific. However, the assessment report acknowledges that recent tagging research suggests that bigeye undertake extensive longitudinal movements, which may be at odds with this assumption. The latest assessment report is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04a-BET-Assessment-of-bigeye-2016.pdf .; This series shows the modeled female spawning stock biomass index from the 2017 stock assessment, which was an update of the 2016 full assessment. Yellowfin tuna are assessed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. They are assumed to comprise one stock throughout the Pacific, although tagging data suggest considerable regional fidelity. In the eastern Pacific, they are primarily fished in tropical waters, from Baja California south. The latest assessment report is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04b-YFT-Assessment-of-yellowfin-2016.pdf.; Estimates of annual recruitment are derived from the stock assessment model. The modeled spawning stock biomass from the latest (2016) stock assessment report was completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Blue marlin are considered to be one stock throughout the Pacific Ocean, and the majority of catch is from pelagic longlines. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; This index represents modeled female spawning stock biomass from the latest (2016) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Blue marlin are considered to be one stock throughout the Pacific Ocean, and the majority of catch is from pelagic longlines. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Annual recruitment is derived from the stock assessment model, and is primarily indexed by catches from troll fisheries on age-0 juvenile fish near Japan. m the latest (2016) stock assessment rep was complted byleted through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Pacific bluefin are considered to be one stock throughout the Pacific Ocean, and are fished throughout their range by multiple countries and fishing gears. At present, the majority are caught by purse seine gear. Their population dynamics are assessed using a fully integrated age-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html .; Estimates of recruitment are derived from the assessment model. Skipjack tuna are assumed to comprise one contiguous stock throughout the Pacific ocean. In the eastern Pacific, they are primarily fished in tropical waters, using purse seine gear. Skipjack are difficult to assess with standard stock assessment methods, due to high and variable productivity, and uncertainties in natural mortality and growth. This species is thus assessed using a simple model which generates indicators of biomass, recruitment and exploitation rate, and compares these to historically observed values (Maunder & Deriso 2007). The stock assessment is completed through the Inter-American Tropical Tuna Commission (IATTC). The relative biomass index shown is from the latest update assessment, including data up to 2016, which is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04c-SKJ-Status-of-skipjack-2016.pdf.; This index represents modeled spawning stock biomass from the latest (2016) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North cific Ocean (ISC). Pacific bluefin are considered to be one stock throughout the Pacific Ocean, and are fished throughout their range by multiple countries and fishing gears. At present, the majority are caught by purse seine gear. Their population dynamics are assessed using a fully integrated age-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Skipjack tuna are assumed to comprise one contiguous stock throughout the Pacific ocean. In the eastern Pacific, they are primarily fished in tropical waters, using purse seine gear. Skipjack are difficult to assess with standard stock assessment methods, due to high and variable productivity, and uncertainties in natural mortality and growth. This species is thus assessed using a simple model which generates indicators of biomass, recruitment and exploitation rate, and compares these to historically observed values (Maunder & Deriso 2007). The stock assessment is completed through the Inter-American Tropical Tuna Commission (IATTC). The relative biomass index shown is from the latest update assessment, including data up to 2016, which is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04c-SKJ-Status-of-skipjack-2016.pdf.; These indices represent modeled exploitable biomass for the western central and eastern Pacific swordfish stocks, from the latest (2014) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Swordfish are considered to comprise two stocks in the North Pacific. The western and central Pacific stock is located throughout the entire North Pacific, except for off Baja California, and central and south America, which is occupied by the eastern Pacific stock. The highest catches are from pelagic longline gears. The population dynamics of both stocks are assessed using a Bayesian state-space generalized surplus production model. The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Estimates of annual recruitment are derived from the stock assessment model.Â The latest (2017) stock assessment report was completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). North Pacific albacore are considered to be one stock throughout the North Pacific Ocean, although some studies have suggested that a northern and southern stock may be present within the assessment area. They are fished throughout their range by multiple countries, mostly using surface gear (troll and pole and line), as well as pelagic longlines and other gears. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html; This index shows modeled female spawning stock biomass of bigeye tuna from the latest stock assessment report, which was completed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. The assessment assumes that there is one stock of bigeye in the eastern Pacific. However, the assessment report acknowledges that recent tagging research suggests that bigeye undertake extensive longitudinal movements, which may be at odds with this assumption. The latest assessment report is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04a-BET-Assessment-of-bigeye-2016.pdf.; Estimates of recruitment are derived from the assessment model. The 2017 stock assessment was an update of the 2016 full assessment. Yellowfin tuna are assessed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. They are assumed to comprise one stock throughout the Pacific, although tagging data suggest considerable regional fidelity. In the eastern Pacific, they are primarily fished in tropical waters, from Baja California south. The latest assessment report is available from http://www.iattc.org/Meetings/Meetings2017/SAC08/PDFs/SAC-08-04b-YFT-Assessment-of-yellowfin-2016.pdf.;

cciea_HMS

Format

A data frame with 2851 rows and 5 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) [-5.99616E8, 1.5778368E9]

albacore_tuna_recruitment

Albacore tuna recruitment (thousands of fish) [112811.0, 256880.0]

albacore_tuna_recruitment_SEup

Albacore tuna recruitment SEup (thousands of fish) [141942.1, 295771.1]

albacore_tuna_recruitment_SElo

Albacore tuna recruitment SElo (thousands of fish) [83679.9, 217988.9]

albacore_tuna_recruitment_SEtype

Albacore tuna recruitment SE type () []

albacore_tuna_spawning_stock_biomass

Albacore tuna spawning stock biomass (metric tonnes) [52465.5, 86714.6]

albacore_tuna_spawning_stock_biomass_SEup

Albacore tuna spawning stock biomass SEup (metric tonnes) [66807.3, 108882.4]

albacore_tuna_spawning_stock_biomass_SElo

Albacore tuna spawning stock biomass SElo (metric tonnes) [38123.7, 64546.8]

albacore_tuna_spawning_stock_biomass_SEtype

Albacore tuna spawning stock biomass SE type () []

bigeye_tuna_recruitment_optimistic

Bigeye tuna recruitment optimistic (thousands of fish) [25361.0, 62830.0]

bigeye_tuna_recruitment_optimistic_SEup

Bigeye tuna recruitment optimistic SEup (thousands of fish) [35741.0, 91464.1]

bigeye_tuna_recruitment_optimistic_SElo

Bigeye tuna recruitment optimistic SElo (thousands of fish) [17645.0, 42846.0]

bigeye_tuna_recruitment_optimistic_SEtype

Bigeye tuna recruitment optimistic SE type () []

bigeye_tuna_recruitment_pessimistic

Bigeye tuna recruitment pessimistic (thousands of fish) [17413.0, 42614.0]

bigeye_tuna_recruitment_pessimistic_SEup

Bigeye tuna recruitment pessimistic SEup (thousands of fish) [20315.0, 52259.9]

bigeye_tuna_recruitment_pessimistic_SElo

Bigeye tuna recruitment pessimistic SElo (thousands of fish) [14327.6, 36075.0]

bigeye_tuna_recruitment_pessimistic_SEtype

Bigeye tuna recruitment pessimistic SE type () []

bigeye_tuna_spawning_stock_biomass_optimistic

Bigeye tuna spawning stock biomass optimistic (metric tonnes) [223595.3, 365256.0]

bigeye_tuna_spawning_stock_biomass_optimistic_SEup

Bigeye tuna spawning stock biomass optimistic SEup (metric tonnes) [405962.0, 668900.0]

bigeye_tuna_spawning_stock_biomass_optimistic_SElo

Bigeye tuna spawning stock biomass optimistic SElo (metric tonnes) [62773.0, 112185.0]

bigeye_tuna_spawning_stock_biomass_optimistic_SEtype

Bigeye tuna spawning stock biomass optimistic SE type () []

bigeye_tuna_spawning_stock_biomass_pessimistic

Bigeye tuna spawning stock biomass pessimistic (metric tonnes) [59567.01, 101975.1]

bigeye_tuna_spawning_stock_biomass_pessimistic_SEup

Bigeye tuna spawning stock biomass pessimistic SEup (metric tonnes) [97979.0, 147028.0]

bigeye_tuna_spawning_stock_biomass_pessimistic_SElo

Bigeye tuna spawning stock biomass pessimistic SElo (metric tonnes) [31691.0, 64157.0]

bigeye_tuna_spawning_stock_biomass_pessimistic_SEtype

Bigeye tuna spawning stock biomass pessimistic SE type () []

bluefin_tuna_recruitment

Bluefin tuna recruitment (thousands of fish) [3194.88, 34812.8]

bluefin_tuna_recruitment_SEup

Bluefin tuna recruitment SEup (thousands of fish) [4149.436, 39464.52]

bluefin_tuna_recruitment_SElo

Bluefin tuna recruitment SElo (thousands of fish) [2240.324, 30194.73]

bluefin_tuna_recruitment_SEtype

Bluefin tuna recruitment SE type () []

bluefin_tuna_spawning_stock_biomass

Bluefin tuna spawning stock biomass (metric tonnes) [10837.1, 156302.0]

bluefin_tuna_spawning_stock_biomass_SEup

Bluefin tuna spawning stock biomass SEup (metric tonnes) [12789.73, 212828.6]

bluefin_tuna_spawning_stock_biomass_SElo

Bluefin tuna spawning stock biomass SElo (metric tonnes) [8884.47, 99775.4]

bluefin_tuna_spawning_stock_biomass_SEtype

Bluefin tuna spawning stock biomass SE type () []

blue_marlin_recruitment

Blue marlin recruitment (thousands of fish) [589.3, 1181.18]

blue_marlin_recruitment_SEup

Blue marlin recruitment SEup (thousands of fish) [705.14, 1399.42]

blue_marlin_recruitment_SElo

Blue marlin recruitment SElo (thousands of fish) [473.46, 964.91]

blue_marlin_recruitment_SEtype

Blue marlin recruitment SE type () []

blue_marlin_spawning_stock_biomass

Blue marlin spawning stock biomass (metric tonnes) [20972.0, 71806.5]

blue_marlin_spawning_stock_biomass_SEup

Blue marlin spawning stock biomass SEup (metric tonnes) [23548.58, 91587.2]

blue_marlin_spawning_stock_biomass_SElo

Blue marlin spawning stock biomass SElo (metric tonnes) [18395.42, 52025.8]

blue_marlin_spawning_stock_biomass_SEtype

Blue marlin spawning stock biomass SE type () []

skipjack_tuna_biomass

Skipjack tuna relative biomass index (metric tonnes) [0.4711875, 2.244901]

skipjack_tuna_biomass_SEup

Skipjack tuna relative biomass index SEup (metric tonnes) []

skipjack_tuna_biomass_SElo

Skipjack tuna relative biomass index SElo (metric tonnes) []

skipjack_tuna_biomass_SEtype

Skipjack tuna relative biomass index SE type () []

skipjack_tuna_recruitment

Skipjack tuna recruitment index (thousands of fish) [0.4760525, 2.503387]

skipjack_tuna_recruitment_SEup

Skipjack tuna recruitment index SEup (thousands of fish) []

skipjack_tuna_recruitment_SElo

Skipjack tuna recruitment index SElo (thousands of fish) []

skipjack_tuna_recruitment_SEtype

Skipjack tuna recruitment index SE type () []

swordfish_eastern_pacific_exploitable_total_biomass

Eastern Pacific exploitable swordfish total biomass (metric tonnes) [31510.0, 67070.0]

swordfish_eastern_pacific_exploitable_total_biomass_SEup

Eastern Pacific exploitable swordfish total biomass SEup (metric tonnes) [42580.0, 89380.0]

swordfish_eastern_pacific_exploitable_total_biomass_SElo

Eastern Pacific exploitable swordfish total biomass SElo (metric tonnes) [20440.0, 44760.0]

swordfish_eastern_pacific_exploitable_total_biomass_SEtype

Eastern Pacific exploitable swordfish total biomass SE type () []

swordfish_western_central_pacific_exploitable_recruitment

Western Central Pacific swordfish recruitment (thousands of fish) [401.0, 1241.0]

swordfish_western_central_pacific_exploitable_recruitment_SEup

Western Central Pacific swordfish recruitment SEup (thousands of fish) [487.0, 1856.0]

swordfish_western_central_pacific_exploitable_recruitment_SElo

Western Central Pacific swordfish recruitment SElo (thousands of fish) [315.0, 949.0]

swordfish_western_central_pacific_exploitable_recruitment_SEtype

Western Central Pacific swordfish recruitment SE type () []

swordfish_western_central_pacific_spawning_stock_biomass

Western Central Pacific swordfish spawning stock biomass (metric tonnes) [17191.0, 44100.0]

swordfish_western_central_pacific_spawning_stock_biomass_SEup

Western Central Pacific swordfish spawning stock biomass SEup (metric tonnes) [21959.0, 52738.0]

swordfish_western_central_pacific_spawning_stock_biomass_SElo

Western Central Pacific swordfish spawning stock biomass SElo (metric tonnes) [12423.0, 35462.0]

swordfish_western_central_pacific_spawning_stock_biomass_SEtype

Western Central Pacific swordfish spawning stock biomass SE type () []

yellowfin_tuna_recruitment

Yellowfin tuna recruitment (thousands of fish) [598155.6, 2035279.0]

yellowfin_tuna_recruitment_SEup

Yellowfin tuna recruitment SEup (thousands of fish) [911885.0, 3308169.0]

yellowfin_tuna_recruitment_SElo

Yellowfin tuna recruitment SElo (thousands of fish) [246326.1, 1244589.0]

yellowfin_tuna_recruitment_SEtype

Yellowfin tuna recruitment SE type () []

yellowfin_tuna_spawning_stock_biomass_index

Yellowfin tuna spawning stock biomass (metric tonnes) [6244.701, 26607.24]

yellowfin_tuna_spawning_stock_biomass_index_SEup

Yellowfin tuna spawning stock biomass SEup (metric tonnes) [13145.74, 66526.3]

yellowfin_tuna_spawning_stock_biomass_index_SElo

Yellowfin tuna spawning stock biomass SElo (metric tonnes) [2892.493, 8664.75]

yellowfin_tuna_spawning_stock_biomass_index_SEtype

Yellowfin tuna spawning stock biomass SE type () []

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HMS/index.html