Chinook Abundance, California

Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement by natural origin Spring-run Chinook spawners to the Central Valley system. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement (based on carcass surveys) by natural origin Winter-run Chinook spawners to the Central Valley system. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement by natural origin Fall Run Chinook to the Klamath and Trinity River systems. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement by natural origin Fall-run Chinook spawners to the Central Valley system, based on the following watersheds: Antelope Creek, Battle Creek, Big Chico Creek, Butte Creek, Clear Creeek, Cottonwood Creek, Deer Creek, Feather River hatchery, and Mill Creek. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement by natural origin Late Fall-run Chinook spawners to the Central Valley system. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement by natural origin Fall-run Chinook to the Klamath and Trinity River systems. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation., Source Data: Various; see Wells et al. 2014, Table S3. Additional Calculations: Abundance indices are calculated as longterm anomalies (observed mean/standard deviation) of annual escapement (live/dead counts) by natural origin spawners from the California Coast, based on the following watersheds: Tomki Creek, Cannon Creek, and Sprowl Creek. Data series for multiple subpopulations were standardized by subtracting the series mean and dividing by the series standard deviation. If a consolidated index for the stock was needed we computed an annual weighted average of the standardized series, with weights proportional to the average abundance for each subpopulation.

cciea_SM_CA_CH_ABND

Format

A data frame with 240 rows and 3 variables:

time

Time (seconds since 1970-01-01T00:00:00Z) 3.1536E7, 1.5147648E9

population

Population () []

abundance_anomaly

Abundance anomaly (Abundance anomaly) -1.671656, 3.423461

Source

https://oceanview.pfeg.noaa.gov/erddap/info/cciea_SM_CA_CH_ABND/index.html