Plot time series, original

This function will plot the time series data. See the initial portion of the function for parts that you can alter.

PlotTimeSeries(
  dframe = NA,
  X = NA,
  Y = NA,
  Y2 = NULL,
  Y.outlier = NULL,
  threshold = NA,
  threshold.loc = "below",
  threshold.correct = FALSE,
  sig.trend = NA,
  sig.mean = NA,
  stats.years = NA,
  LWD = 1,
  LWD.threshold = 1,
  SE = NULL,
  SEup = NA,
  SElo = NA,
  XLIM = NA,
  YLIM = NA,
  Xlab = NA,
  Ylab = NULL,
  Title = NULL,
  Xangle = 0,
  Xaxis.cex = 0.9,
  Yaxis.cex = 0.9,
  Ylab.cex = 0.9,
  TitleCex = 0.9,
  YATS = 3,
  Y.lab.drop = FALSE,
  Y.axis.labels = NA,
  yminor = NA,
  Label.cex = 0.9,
  Pt.cex = 0.3,
  TicsEvery = 10,
  MinorTics = 1,
  ALPHA = 200,
  PERIOD = 5,
  Arrow.cex = 2,
  AltTxt.Period = "5-years",
  AltTxt = NA,
  AltTxt.file.name = NA,
  Mar = c(0, 3.5, 0.7, 0),
  Bmar = NA,
  Lmar = NA,
  Tmar = NA,
  Rmar = NA,
  Ylab_line = 2.5
)

Arguments

dframe

expected column names are: "year", "index", "timeseries", "metric","SEup", "SElo", "SE"

X

column name, typically "year", to plot on x axis

Y

column name, typically "index", to plot on y axis

Y2

...

Y.outlier

...

threshold

threshold value

threshold.loc

...

threshold.correct

...

sig.trend

significant trend

sig.mean

significant mean

stats.years

...

LWD

...

LWD.threshold

...

SE

...

SEup

NOTE: if you call SEup separately (say to plot SD instead of SE), the first value cannot be an NA. Set it and SElo to the index data point for that value.

SElo

...

XLIM

...

YLIM

...

Xlab

...

Ylab

...

Title

...

Xangle

...

Xaxis.cex

...

Yaxis.cex

...

Ylab.cex

...

TitleCex

...

YATS

...

Y.lab.drop

...

Y.axis.labels

...

yminor

...

Label.cex

...

Pt.cex

...

TicsEvery

...

MinorTics

...

ALPHA

...

PERIOD

...

Arrow.cex

...

AltTxt.Period

...

AltTxt

...

AltTxt.file.name

...

Mar

...

Bmar

...

Lmar

...

Tmar

...

Rmar

...

Ylab_line

...

Value

Details

The default plot is the standard IEA time series plot. If you include a "treshold" value, the plot will shift format to plotting versus a threshold. If so, threshold.loc, will swap the plotting of grey polygons to indicate periods where the timeseries is above or below the threshold.

There are special plots for Seabirds: For seabird plots you must include both Y2 and Y.outlier in the plot call. For seabird mortality, Y is all the data, Y2 is data with outliers removed, Y.outlier are just the outliers. Trends are evaluated with the outliers removed.