Plot time series (original)

This function will plot the time series data. See the initial portion of the function for parts that you can alter.

PlotTimeSeries(
  dframe = NA,
  X = NA,
  Y = NA,
  Y2 = NULL,
  Y.outlier = NULL,
  threshold = NA,
  threshold.loc = "below",
  threshold.correct = FALSE,
  sig.trend = NA,
  sig.mean = NA,
  stats.years = NA,
  LWD = 1,
  LWD.threshold = 1,
  SE = NULL,
  SEup = NA,
  SElo = NA,
  XLIM = NA,
  YLIM = NA,
  Xlab = NA,
  Ylab = NULL,
  Title = NULL,
  Xangle = 0,
  Xaxis.cex = 0.9,
  Yaxis.cex = 0.9,
  Ylab.cex = 0.9,
  TitleCex = 0.9,
  YATS = 3,
  Y.lab.drop = FALSE,
  Y.axis.labels = NA,
  yminor = NA,
  Label.cex = 0.9,
  Pt.cex = 0.3,
  TicsEvery = 10,
  MinorTics = 1,
  ALPHA = 200,
  PERIOD = 5,
  Arrow.cex = 2,
  AltTxt.Period = "5-years",
  AltTxt = NA,
  AltTxt.file.name = NA,
  Mar = c(0, 3.5, 0.7, 0),
  Bmar = NA,
  Lmar = NA,
  Tmar = NA,
  Rmar = NA,
  Ylab_line = 2.5
)

Arguments

dframe
expected column names are: "year", "index", "timeseries", "metric","SEup", "SElo", "SE"

X
column name, typically "year", to plot on x axis

Y
column name, typically "index", to plot on y axis

Y2
...

Y.outlier
...

threshold
threshold value

threshold.loc
...

threshold.correct
...

sig.trend
significant trend

sig.mean
significant mean

stats.years
...

LWD
...

LWD.threshold
...

SE
...

SEup
NOTE: if you call SEup separately (say to plot SD instead of SE), the first value cannot be an NA. Set it and SElo to the index data point for that value.

SElo
...

XLIM
...

YLIM
...

Xlab
...

Ylab
...

Title
...

Xangle
...

Xaxis.cex
...

Yaxis.cex
...

Ylab.cex
...

TitleCex
...

YATS
...

Y.lab.drop
...

Y.axis.labels
...

yminor
...

Label.cex
...

Pt.cex
...

TicsEvery
...

MinorTics
...

ALPHA
...

PERIOD
...

Arrow.cex
...

AltTxt.Period
...

AltTxt
...

AltTxt.file.name
...

Mar
...

Bmar
...

Lmar
...

Tmar
...

Rmar
...

Ylab_line
...

Value

Details

The default plot is the standard IEA time series plot. If you include a "treshold" value, the plot will shift format to plotting versus a threshold. If so, threshold.loc, will swap the plotting of grey polygons to indicate periods where the timeseries is above or below the threshold.

There are special plots for Seabirds: For seabird plots you must include both Y2 and Y.outlier in the plot call. For seabird mortality, Y is all the data, Y2 is data with outliers removed, Y.outlier are just the outliers. Trends are evaluated with the outliers removed.

Examples

PlotTimeSeries

#> function (dframe = NA, X = NA, Y = NA, Y2 = NULL, Y.outlier = NULL, 
#>     threshold = NA, threshold.loc = "below", threshold.correct = FALSE, 
#>     sig.trend = NA, sig.mean = NA, stats.years = NA, LWD = 1, 
#>     LWD.threshold = 1, SE = NULL, SEup = NA, SElo = NA, XLIM = NA, 
#>     YLIM = NA, Xlab = NA, Ylab = NULL, Title = NULL, Xangle = 0, 
#>     Xaxis.cex = 0.9, Yaxis.cex = 0.9, Ylab.cex = 0.9, TitleCex = 0.9, 
#>     YATS = 3, Y.lab.drop = FALSE, Y.axis.labels = NA, yminor = NA, 
#>     Label.cex = 0.9, Pt.cex = 0.3, TicsEvery = 10, MinorTics = 1, 
#>     ALPHA = 200, PERIOD = 5, Arrow.cex = 2, AltTxt.Period = "5-years", 
#>     AltTxt = NA, AltTxt.file.name = NA, Mar = c(0, 3.5, 0.7, 
#>         0), Bmar = NA, Lmar = NA, Tmar = NA, Rmar = NA, Ylab_line = 2.5) 
#> {
#>     mn.col = "black"
#>     sd.col = box.col = threshold.col = rgb(63, 173, 213, max = 255)
#>     pt.col = rgb(225, 118, 44, max = 255)
#>     line.col = "black"
#>     error.col = rgb(211, 211, 211, alpha = ALPHA, max = 255)
#>     arrow.color = "black"
#>     threshold.fill.col = "lightgrey"
#>     title.font = 2
#>     if (is.na(Bmar == TRUE)) {
#>         Bmar = Mar[1]
#>     }
#>     if (is.na(Lmar == TRUE)) {
#>         Lmar = Mar[2]
#>     }
#>     if (is.na(Tmar == TRUE)) {
#>         Tmar = Mar[3]
#>     }
#>     if (is.na(Rmar == TRUE)) {
#>         Rmar = Mar[4]
#>     }
#>     par(ps = 10, cex = 1, cex.axis = Yaxis.cex, cex.main = TitleCex, 
#>         cex.lab = Label.cex, mar = c(Bmar, Lmar, Tmar, Rmar))
#>     if (length(dframe) == 1) {
#>         X = X
#>         Y = Y
#>         TS = Title
#>     }
#>     if (length(dframe) != 1) {
#>         X = dframe$year
#>         Y = dframe$index
#>         if (is.na(Y2) || is.null(Y2)) {
#>             Y2 <- Y
#>         }
#>         TS = Title
#>         if (is.null(Title)) {
#>             Title = dframe$timeseries[1]
#>             TS = Title
#>         }
#>         else {
#>             if (is.na(Title)) {
#>                 Title = NA
#>                 TS = dframe$timeseries[1]
#>             }
#>         }
#>         if (is.null(Ylab)) {
#>             Ylab = dframe$metric[1]
#>         }
#>         else {
#>             if (is.na(Ylab)) {
#>                 Ylab = NA
#>             }
#>         }
#>         if (is.na(SEup[1] == TRUE)) {
#>             if (length(SEup) > 1) {
#>                 SEup[1] = SElo[1] <- Y[1]
#>             }
#>         }
#>         if (is.na(SEup[1]) == TRUE) {
#>             cn = colnames(dframe)
#>             se.up = grep("SEup", cn)
#>             if (length(se.up) > 0) {
#>                 SEup = dframe$SEup
#>                 SElo = dframe$SElo
#>             }
#>             if (is.null(SE)) {
#>                 if (length(se.up) == 0) {
#>                   if (!is.null(dframe$SE)) {
#>                     SEup = dframe$index + dframe$SE
#>                     SElo = dframe$index - dframe$SE
#>                   }
#>                 }
#>                 if (length(se.up) == 0) {
#>                   if (!is.null(dframe$se)) {
#>                     SEup = dframe$index + dframe$se
#>                     SElo = dframe$index - dframe$se
#>                   }
#>                 }
#>             }
#>         }
#>     }
#>     if (nchar(as.character(X[1])) > 10) {
#>         X = substring(X, 1, 10)
#>     }
#>     if (nchar(as.character(X[1])) == 10) {
#>         x = as.POSIXct(X)
#>     }
#>     if (nchar(as.character(X[1])) == 7) {
#>         x = as.POSIXct(paste(X, 1, sep = "-"))
#>     }
#>     if (nchar(as.character(X[1])) == 4) {
#>         x = as.POSIXct(paste(X, 1, 1, sep = "-"))
#>     }
#>     XY = data.frame(cbind(x, Y))
#>     XXYY = XY
#>     XY = na.omit(XY)
#>     XLIM1 = as.numeric(substring(as.POSIXct(strptime(XLIM[1], 
#>         "%Y")), 1, 4))
#>     XLIM2 = as.numeric(substring(as.POSIXct(strptime(XLIM[2], 
#>         "%Y")), 1, 4))
#>     x1 = as.numeric(substring(as.POSIXct(strptime(x, "%Y")), 
#>         1, 4))
#>     xmin1 = min(x1)
#>     xmin2 = ifelse(is.na(XLIM1) == TRUE, xmin1, XLIM1)
#>     Xmin = as.POSIXct(paste((xmin2 - 1), 11, 1, sep = "-"))
#>     xmax = max(x)
#>     xmax1 = as.POSIXct(paste(max(x1), 1, 1, sep = "-"))
#>     if (xmax > xmax1) {
#>         xmax2 = max(x1) + 1
#>     }
#>     else {
#>         xmax2 = max(x1)
#>     }
#>     xmax3 = ceiling(xmin2 + 1.02 * (xmax2 - xmin2))
#>     xmax4 = ifelse(is.na(XLIM2) == TRUE, xmax3, ceiling(min(xmin2) + 
#>         1.02 * (XLIM2 - xmin2)))
#>     Xmax = as.POSIXct(paste(xmax4, 1, 1, sep = "-"))
#>     xL = length(X)
#>     xmax5 = ifelse(is.na(XLIM2) == TRUE, xmax2, XLIM2)
#>     labrange = c(xmin2:xmax5)
#>     atx = as.POSIXct(paste(labrange, 1, 1, sep = "-"))
#>     xlabs = as.numeric(substring(as.POSIXct(strptime(atx, "%Y")), 
#>         1, 4))
#>     Y[Y %in% c(-999, -9999)] <- NA
#>     if (!is.null(Y2)) {
#>         Y2[Y2 %in% c(-999, -9999)] <- NA
#>     }
#>     SEup[SEup %in% c(-999, -9999)] <- NA
#>     SElo[SElo %in% c(-999, -9999)] <- NA
#>     if (!is.null(Y.outlier)) {
#>         Y.calc = Y2
#>     }
#>     else {
#>         Y.calc = Y
#>     }
#>     if (is.na(stats.years[1]) == TRUE) {
#>         stdev = sd(Y.calc, na.rm = TRUE)
#>         mn = mean(Y.calc, na.rm = TRUE)
#>         Usd = mn + stdev
#>         Lsd = mn - stdev
#>         stdev.plot = sd(Y, na.rm = TRUE)
#>         mn.plot = mean(Y, na.rm = TRUE)
#>         Usd.plot = mn.plot + stdev.plot
#>         Lsd.plot = mn.plot - stdev.plot
#>     }
#>     else {
#>         x.yr = substr(x, 1, 4)
#>         Y.temp = data.frame(cbind(x.yr, Y.calc))
#>         Y.calc2 = as.numeric(as.character(Y.temp[Y.temp$x %in% 
#>             stats.years, "Y.calc"]))
#>         stdev = sd(Y.calc2, na.rm = TRUE)
#>         mn = mean(Y.calc2, na.rm = TRUE)
#>         Usd = mn + stdev
#>         Lsd = mn - stdev
#>     }
#>     if (is.na(YLIM[1]) == TRUE) {
#>         ymin = min(Y, Y2, SElo, Lsd, na.rm = TRUE)
#>     }
#>     else {
#>         ymin = YLIM[1]
#>     }
#>     if (length(YLIM == 2)) {
#>         if (is.na(YLIM[2]) == TRUE) {
#>             ymax = max(Y, Y2, SEup, Usd, na.rm = TRUE)
#>         }
#>         else {
#>             ymax = YLIM[2]
#>         }
#>     }
#>     if (is.na(Y.axis.labels[1]) == TRUE) {
#>         yats = pretty(c(ymin, ymax), n = YATS)
#>     }
#>     else {
#>         yats = Y.axis.labels
#>     }
#>     maxats = max(yats)
#>     Ymin1 = ifelse(is.na(YLIM[1]) == TRUE, ymin, YLIM[1])
#>     Ymax = ifelse(is.na(YLIM[2]) == TRUE, max(ymax, maxats), 
#>         YLIM[2])
#>     Ymin = Ymin1 - 1 * (Ymax - Ymin1)
#>     mdfy = length(xlabs)/100
#>     xshift = max(mdfy * diff(xlabs)[1], 0.3)
#>     plot(Y ~ x, xlab = NA, ylab = NA, pch = "19", type = "n", 
#>         xlim = c(Xmin, Xmax), ylim = c(Ymin, Ymax), bty = "n", 
#>         xaxt = "n", yaxt = "n")
#>     xline = min(min(yats) - 0.05 * (par()$usr[4] - par()$usr[3]), 
#>         ymin)
#>     modif1 = ifelse(length(atx) > 21, 0.06, 0.05)
#>     tl = min(xline) - modif1 * (par()$usr[4] - par()$usr[3])
#>     tl2 = min(xline) - 0.03 * (par()$usr[4] - par()$usr[3])
#>     modif2 = ifelse(length(atx) > 21, 0.16, 0.14)
#>     txtloc = min(xline) - modif2 * (par()$usr[4] - par()$usr[3])
#>     if (Y.lab.drop == TRUE) {
#>         g = seq(1, length(yats), 2)
#>         ymajor = yats[g]
#>         h = seq(2, length(yats), 2)
#>         yminor = yats[h]
#>     }
#>     else {
#>         ymajor <- yats
#>     }
#>     yats.label = format(ymajor)
#>     axis(side = 2, at = ymajor, labels = yats.label, las = 1, 
#>         cex = 1)
#>     if (is.na(yminor[1]) == FALSE) {
#>         axis(side = 2, at = yminor, labels = NA, tck = -0.05)
#>     }
#>     if (is.na(TicsEvery) == TRUE) {
#>         if (Xangle == 0) {
#>             if (length(atx) <= 13) {
#>                 lines(c(Xmin, atx[length(atx)]), c(xline, xline))
#>                 for (g in 1:length(atx)) {
#>                   arrows(atx[g], xline, atx[g], tl, length = 0)
#>                 }
#>                 for (g in 1:length(atx)) {
#>                   text(atx[g], txtloc, xlabs[g], srt = Xangle, 
#>                     cex = Xaxis.cex)
#>                 }
#>             }
#>             if (length(atx) > 13) {
#>                 seq.num = 3
#>                 if ((length(atx)/2 == floor(length(atx)/2)) == 
#>                   TRUE) {
#>                   seq.num = 2
#>                 }
#>                 if ((length(atx)/3 == floor(length(atx)/3)) == 
#>                   TRUE) {
#>                   seq.num = 3
#>                 }
#>                 if ((length(atx)/4 == floor(length(atx)/4)) == 
#>                   TRUE) {
#>                   seq.num = 4
#>                 }
#>                 if (length(atx) > 50) {
#>                   seq.num = 5
#>                 }
#>                 if (length(atx) > 99) {
#>                   seq.num = 10
#>                 }
#>                 atx2 = as.numeric(substring(as.character(atx), 
#>                   1, 4))
#>                 atx3 = rev(abs(seq(-max(atx2), -min(atx2), seq.num)))
#>                 atx4 = as.POSIXct(paste(atx3, 1, 1, sep = "-"))
#>                 lines(c(Xmin, atx[length(atx)]), c(xline, xline))
#>                 for (g in 1:length(atx)) {
#>                   arrows(atx4[g], xline, atx4[g], tl, length = 0)
#>                 }
#>                 for (g in 1:length(atx3)) {
#>                   text(atx4[g], txtloc, atx3[g], srt = Xangle, 
#>                     cex = Xaxis.cex)
#>                 }
#>             }
#>         }
#>         if (Xangle != 0) {
#>             if (length(atx) <= 13) {
#>                 lines(c(Xmin, atx[length(atx)]), c(xline, xline))
#>                 for (g in 1:length(atx)) {
#>                   arrows(atx[g], xline, atx[g], tl, length = 0)
#>                 }
#>                 for (g in 1:length(atx)) {
#>                   text(atx[g] - xshift, txtloc, xlabs[g], srt = Xangle, 
#>                     cex = Xaxis.cex)
#>                 }
#>             }
#>             if (length(atx) > 13) {
#>                 seq.num = 3
#>                 if ((length(atx)/2 == floor(length(atx)/2)) == 
#>                   TRUE) {
#>                   seq.num = 2
#>                 }
#>                 if ((length(atx)/3 == floor(length(atx)/3)) == 
#>                   TRUE) {
#>                   seq.num = 3
#>                 }
#>                 if ((length(atx)/4 == floor(length(atx)/4)) == 
#>                   TRUE) {
#>                   seq.num = 4
#>                 }
#>                 if (length(atx) > 50) {
#>                   seq.num = 5
#>                 }
#>                 if (length(atx) > 99) {
#>                   seq.num = 10
#>                 }
#>                 atx2 = as.numeric(substring(as.character(atx), 
#>                   1, 4))
#>                 atx3 = rev(abs(seq(-max(atx2), -min(atx2), seq.num)))
#>                 atx4 = as.POSIXct(paste(atx3, 1, 1, sep = "-"))
#>                 lines(c(Xmin, atx[length(atx)]), c(xline, xline))
#>                 for (g in 1:length(atx)) {
#>                   arrows(atx4[g], xline, atx4[g], tl, length = 0)
#>                 }
#>                 for (g in 1:length(atx3)) {
#>                   text(atx4[g] - xshift, txtloc, atx3[g], srt = Xangle, 
#>                     cex = Xaxis.cex)
#>                 }
#>             }
#>         }
#>     }
#>     else {
#>         atx2 = as.numeric(substring(as.character(atx), 1, 4))
#>         atx3 = rev(abs(seq(-max(atx2), -min(atx2), TicsEvery)))
#>         atx4 = as.POSIXct(paste(atx3, 1, 1, sep = "-"))
#>         lines(c(Xmin, atx[length(atx)]), c(xline, xline))
#>         for (g in 1:length(atx)) {
#>             arrows(atx4[g], xline, atx4[g], tl, length = 0)
#>         }
#>         for (g in 1:length(atx3)) {
#>             text(atx4[g], txtloc, atx3[g], srt = Xangle, cex = Xaxis.cex)
#>         }
#>     }
#>     if (is.na(MinorTics) == TRUE) {
#>         atx2 = as.numeric(substring(as.character(atx), 1, 4))
#>         if (length(atx) > 99) {
#>             atx5 = rev(abs(seq(-max(atx2), -min(atx2), 5)))
#>         }
#>         else {
#>             atx5 = atx
#>         }
#>         atx6 = as.POSIXct(paste(atx5, 1, 1, sep = "-"))
#>         for (g in 1:length(atx5)) {
#>             arrows(atx6[g], xline, atx6[g], tl2, length = 0)
#>         }
#>     }
#>     else {
#>         atx2 = as.numeric(substring(as.character(atx), 1, 4))
#>         atx5 = rev(abs(seq(-max(atx2), -min(atx2), MinorTics)))
#>         atx6 = as.POSIXct(paste(atx5, 1, 1, sep = "-"))
#>         for (g in 1:length(atx5)) {
#>             arrows(atx6[g], xline, atx6[g], tl2, length = 0)
#>         }
#>     }
#>     period = PERIOD - 1
#>     xeval = atx[(length(atx) - period)]
#>     yloc = length(Y) - period
#>     if (is.null(Y.outlier)) {
#>         Y5 = Y[yloc:length(Y)]
#>     }
#>     else {
#>         Y5 = Y2[yloc:length(Y2)]
#>     }
#>     X5 = x[yloc:length(x)]
#>     x2 = ifelse(is.na(Y), NA, x)[is.na(Y) == FALSE]
#>     y2 = ifelse(is.na(Y), NA, Y)[is.na(Y) == FALSE]
#>     SEup2 = ifelse(is.na(Y), NA, SEup)[is.na(Y) == FALSE]
#>     SElo2 = ifelse(is.na(Y), NA, SElo)[is.na(Y) == FALSE]
#>     if (PERIOD != 0) {
#>         if (is.na(threshold)) {
#>             if (is.null(Y.outlier)) {
#>                 polygon(x = c(X5[1], max(X5), max(X5), X5[1]), 
#>                   y = c(Usd, Usd, Lsd, Lsd), border = NA, col = box.col)
#>                 segments(Xmin, mn, max(x), mn, col = mn.col, 
#>                   lty = "dotted", lwd = 1)
#>                 segments(Xmin, Usd, max(x), Usd, col = sd.col, 
#>                   lwd = 1)
#>                 segments(Xmin, Lsd, max(x), Lsd, col = sd.col, 
#>                   lwd = 1)
#>                 if (is.na(SEup[1]) == FALSE) {
#>                   polygon(c(x2, rev(x2)), c(SEup2, rev(SElo2)), 
#>                     border = NA, col = error.col)
#>                 }
#>             }
#>             if (!is.null(Y.outlier)) {
#>                 polygon(x = c(X5[1], max(X5), max(X5), X5[1]), 
#>                   y = c(Usd.plot, Usd.plot, Lsd.plot, Lsd.plot), 
#>                   col = box.col, border = NA, lwd = 0.5)
#>                 segments(Xmin, Usd, max(x), Usd, lty = "dotted")
#>                 segments(Xmin, Lsd, max(x), Lsd, lty = "dotted")
#>             }
#>             xx <- par()
#>             x1 <- xx$usr[1]
#>             x2 <- xx$usr[2]
#>             m1 = lm(Y5 ~ X5)
#>             s1 <- summary(m1)
#>             b1 <- s1$coefficients[2, 1]
#>             pval <- s1$coefficients[2, 4]
#>             pred = predict(m1)
#>             delta = pred[length(pred)] - pred[1]
#>             Z = abs(delta) - (stdev)
#>             x01 <- (x2 - x1) * 0.97 + x1
#>             y01 = yats[1] + (max(yats) - yats[1]) * 0.7
#>             y02 = yats[1] + (max(yats) - yats[1]) * 0.3
#>             LineWT = 3
#>             afont = 2
#>             arrow.cex = Arrow.cex
#>             if (is.na(sig.trend[1]) == FALSE) {
#>                 Z = ifelse(sig.trend[1] %in% c("positive", "negative"), 
#>                   1, 0)
#>                 delta = ifelse(sig.trend[1] == "positive", 1, 
#>                   -1)
#>             }
#>             if (Z <= 0) {
#>                 text(x01, y01, "\\->", family = "HersheySymbol", 
#>                   cex = arrow.cex, col = arrow.color, font = 2)
#>             }
#>             if (Z > 0 & delta > 0) {
#>                 text(x01, y01, "\\#H2296", family = "HersheySymbol", 
#>                   cex = arrow.cex, col = arrow.color, font = 2)
#>             }
#>             if (Z > 0 & delta < 0) {
#>                 text(x01, y01, "\\#H2299", family = "HersheySymbol", 
#>                   cex = arrow.cex, col = arrow.color, font = 2)
#>             }
#>             mn5 = mean(Y5, na.rm = TRUE)
#>             dMean = mn - mn5
#>             Z2 = abs(dMean) - stdev
#>             if (is.na(sig.mean[1]) == FALSE) {
#>                 Z2 = ifelse(sig.mean[1] %in% c("positive", "negative"), 
#>                   1, 0)
#>                 dMean = ifelse(sig.mean[1] == "positive", -1, 
#>                   1)
#>             }
#>             if (Z2 <= 0) {
#>                 points(x01, y02, cex = 1, pch = 19, col = "black")
#>             }
#>             if (Z2 > 0 & dMean > 0) {
#>                 text(x01, y02, "_", cex = 1, col = arrow.color, 
#>                   font = afont)
#>             }
#>             if (Z2 > 0 & dMean < 0) {
#>                 text(x01, y02, "+", cex = 1, col = arrow.color, 
#>                   font = afont)
#>             }
#>         }
#>     }
#>     XY = data.frame(cbind(x, Y))
#>     XY0 = XY
#>     if (threshold.loc == "above") {
#>         XY0[is.na(XY)] <- 0
#>     }
#>     else {
#>         XY0[is.na(XY)] <- threshold
#>     }
#>     XY = na.omit(XY)
#>     if (!is.na(threshold)) {
#>         if (threshold.loc == "n") {
#>             segments(Xmin, threshold, max(x), threshold, col = threshold.col, 
#>                 lwd = 1)
#>         }
#>         else {
#>             if (threshold.correct == TRUE) {
#>                 d1 = c(XY0$x, rev(XY0$x))
#>                 d2 = c(XY0$Y, rep(threshold, length(XY0$Y)))
#>             }
#>             else {
#>                 d1 = c(XY$x, rev(XY$x))
#>                 d2 = c(XY$Y, rep(threshold, length(XY$Y)))
#>             }
#>             polygon(d1, d2, border = NA, col = threshold.fill.col)
#>             d3.l = length(d1)/2
#>             fill.loc = ifelse(threshold.loc == "above", min(yats), 
#>                 max(yats))
#>             d3 = c(rep(threshold, d3.l), rep(fill.loc, d3.l))
#>             polygon(d1, d3, border = NA, col = "white")
#>             if (threshold > xline) {
#>                 segments(Xmin, threshold, max(x), threshold, 
#>                   col = threshold.col, lwd = LWD.threshold)
#>             }
#>         }
#>     }
#>     if (is.na(LWD) == TRUE) {
#>         LWD = ifelse(100/xL > 2, 0.5, ifelse(100/xL < 1, 0.5, 
#>             1))
#>     }
#>     else {
#>         LWD = LWD
#>     }
#>     if (threshold.correct == TRUE) {
#>         lines(Y ~ x, lwd = LWD, col = line.col)
#>         lines(XY0$Y ~ XY0$x, lwd = LWD, col = line.col, lty = "dotted")
#>     }
#>     else {
#>         lines(Y ~ x, lwd = LWD, col = line.col)
#>         lines(XY$Y ~ XY$x, lwd = LWD, col = line.col, lty = "dotted")
#>     }
#>     Y3 = na.omit(Y2)
#>     if (!is.na(Y3[1])) {
#>         points(Y2 ~ x, pch = 20, col = pt.col, cex = Pt.cex)
#>     }
#>     if (!is.null(Y.outlier)) {
#>         Y.out = data.frame(cbind(x, Y.outlier))
#>         Y.out = na.omit(Y.out)
#>         points(Y.out[, 1], Y.out[, 2], pch = 21, cex = Pt.cex * 
#>             1.5, bg = "white")
#>     }
#>     if (!is.na(threshold)) {
#>         if (is.na(SEup[1]) == FALSE) {
#>             lines(x, SEup, lty = "dotted")
#>             lines(x, SElo, lty = "dotted")
#>         }
#>     }
#>     if (is.na(Xlab) != TRUE) {
#>         mtext(side = 1, Xlab, line = -1)
#>     }
#>     loc = par()$usr[3] + 0.7 * (par()$usr[4] - par()$usr[3])
#>     mtext(side = 2, Ylab, line = Ylab_line, at = loc, cex = Ylab.cex)
#>     if (!is.na(Title) == TRUE) {
#>         title(Title, line = 0, font = title.font, cex.main = TitleCex)
#>     }
#>     if (exists("note.update.status") == TRUE) {
#>         if (note.update.status == TRUE) {
#>             if (dframe$type[1] != "current.data") {
#>                 legend("topleft", legend = "old data", bg = "red")
#>             }
#>             else {
#>                 legend("topleft", legend = "new data", bg = "green")
#>             }
#>         }
#>     }
#>     if (!is.na(AltTxt)) {
#>         d1 = substring(AltTxt, nchar(AltTxt), nchar(AltTxt))
#>         if (d1 != "/") {
#>             AltTxt = paste(AltTxt, "/", sep = "")
#>         }
#>         atxmin = min(X)
#>         atxmax = max(X)
#>         YMin = round(Ymin, 2)
#>         YMax = round(Ymax, 2)
#>         atymin = ifelse(Ymin < 0, paste("negative", abs(YMin)), 
#>             YMin)
#>         atymax = ifelse(Ymax < 0, paste("negative", abs(YMax)), 
#>             YMax)
#>         xy = data.frame(cbind(X, Y))
#>         xy$X = as.character(xy$X)
#>         xy$Y = as.numeric(as.character(xy$Y))
#>         ymx = xy[match(max(na.omit(xy$Y)), xy$Y), ]
#>         ymn = xy[match(min(na.omit(xy$Y)), xy$Y), ]
#>         if (is.null(Ylab)) {
#>             Ylab = dframe$metric[1]
#>         }
#>         if (is.na(Ylab)) {
#>             Ylab = dframe$metric[1]
#>         }
#>         if (is.na(threshold)) {
#>             OpeningStatement = paste0("This figure shows time-series data for ", 
#>                 TS, " (y-axis, ", Ylab, ") from ", atxmin, " to ", 
#>                 atxmax, " (x-axis, Year). ")
#>             Mean_SD = paste0("The mean of the full time series was ", 
#>                 round(mn, 2), " (lower standard deviation = ", 
#>                 round(Lsd, 2), ", upper standard deviation = ", 
#>                 round(Usd, 2), "). ")
#>             HighLowYears = paste0("The index was highest in ", 
#>                 ymx[1, 1], " (at ", round(ymx[1, 2], 2), "), and lowest in ", 
#>                 ymn[1, 1], " (at ", round(ymn[1, 2], 2), "). ")
#>             if (Z <= 0) {
#>                 PeriodTrend = paste0("The index changed by less than one standard deviation of the full time series over the last ", 
#>                   AltTxt.Period, ". ")
#>             }
#>             if (Z > 0 & delta > 0) {
#>                 PeriodTrend = paste0("The index increased by more than one standard deviation of the full time series over the last ", 
#>                   AltTxt.Period, ". ")
#>             }
#>             if (Z > 0 & delta < 0) {
#>                 PeriodTrend = paste0("The index decreased by more than one standard deviation of the full time series over the last ", 
#>                   AltTxt.Period, ". ")
#>             }
#>             if (Z2 <= 0) {
#>                 points(x01, y02, cex = 1, pch = 19)
#>                 PeriodMean = paste0("The mean of the last ", 
#>                   AltTxt.Period, " was within one standard deviation of the long-term mean. ")
#>             }
#>             if (Z2 > 0 & dMean > 0) {
#>                 PeriodMean = paste0("The mean of the last ", 
#>                   AltTxt.Period, "  was more than one standard deviation below the mean of the full time series. ")
#>             }
#>             if (Z2 > 0 & dMean < 0) {
#>                 PeriodMean = paste0("The mean of the last ", 
#>                   AltTxt.Period, " was more than one standard deviation above the mean of the full time series. ")
#>             }
#>             above.sd = na.omit(xy[xy$Y > Usd, ])
#>             if (nrow(above.sd) == 0) {
#>                 AboveYears = "The index was never above one standard deviation of the full time series. "
#>             }
#>             if (nrow(above.sd) == 1) {
#>                 above.last = above.sd[1, 1]
#>                 AboveYears = paste0("The index was more than one standard deviation above the mean of the full time series in ", 
#>                   above.last, ". ")
#>             }
#>             if (nrow(above.sd) == 2) {
#>                 AboveYears = paste0("The index was  more than one standard deviation above the mean of the full time series in ", 
#>                   above.sd[1, 1], " and ", above.sd[2, 1], ". ")
#>             }
#>             if (nrow(above.sd) > 2) {
#>                 no.yrs = nrow(above.sd)
#>                 above.first = above.sd[1, 1]
#>                 above.last = above.sd[no.yrs, 1]
#>                 above = above.first
#>                 for (i in 2:(no.yrs - 1)) {
#>                   above = paste0(above, ", ", above.sd[i, 1])
#>                 }
#>                 above.years = paste0(above, " and ", above.last)
#>                 AboveYears = paste0("The index was more than one standard deviation above the mean of the full time series in ", 
#>                   above.years, ". ")
#>             }
#>             below.sd = na.omit(xy[xy$Y < Lsd, ])
#>             if (nrow(below.sd) == 0) {
#>                 BelowYears = "The index was never below one standard deviation of the full time series. "
#>             }
#>             if (nrow(below.sd) == 1) {
#>                 below.last = below.sd[1, 1]
#>                 BelowYears = paste0("The index was more than one standard deviation below the mean of the full time series in ", 
#>                   below.last, ". ")
#>             }
#>             if (nrow(below.sd) == 2) {
#>                 BelowYears = paste0("The index was more than one standard deviation below the mean of the full time series in ", 
#>                   below.sd[1, 1], " and ", below.sd[2, 1], ". ")
#>             }
#>             if (nrow(below.sd) > 2) {
#>                 no.yrs = nrow(below.sd)
#>                 below.first = below.sd[1, 1]
#>                 below.last = below.sd[no.yrs, 1]
#>                 below = below.first
#>                 for (i in 2:(no.yrs - 1)) {
#>                   below = paste0(below, ", ", below.sd[i, 1])
#>                 }
#>                 below.years = paste0(below, " and ", below.last)
#>                 BelowYears = paste0("The index was more than one standard deviation below the mean of the full time series in ", 
#>                   below.years, ". ")
#>             }
#>             LastYear = paste0("The index was ", round(Y[length(Y)], 
#>                 2), " in ", X[length(X)], ", the most recent year of the time series. ")
#>             AltText = paste0(OpeningStatement, Mean_SD, AboveYears, 
#>                 BelowYears, LastYear)
#>         }
#>         if (!is.na(threshold)) {
#>             OpeningStatement = paste0("This figure shows time-series data for ", 
#>                 TS, " from ", atxmin, " to ", atxmax, " (x-axis, Year). The y-axis (", 
#>                 Ylab, ") varies from ", atymin, " to ", atymax, 
#>                 ". The index is plotted relative to a threshold value of ", 
#>                 threshold, ". The threshold is indicated by a blue line.")
#>             at3 = paste0("The index was highest in ", ymx[1, 
#>                 1], " (at ", round(ymx[1, 2], 2), "), and lowest in ", 
#>                 ymn[1, 1], " (at ", round(ymn[1, 2], 2), "). ", 
#>                 sep = "")
#>             above.th = na.omit(xy[xy$Y > threshold, ])
#>             if (nrow(above.th) == 0) {
#>                 AboveYears = paste0("The index was never above the threshold")
#>             }
#>             if (nrow(above.th) == 1) {
#>                 above.last = above.th[1, 1]
#>                 AboveYears = paste0("The index was above the threshold value of")
#>             }
#>             if (nrow(above.th) == 2) {
#>                 AboveYears = paste0("The index was above the threshold in ", 
#>                   above.th[1, 1], " and ", above.th[2, 1], ". ")
#>             }
#>             if (nrow(above.th) > 2) {
#>                 no.yrs = nrow(above.th)
#>                 above.first = above.th[1, 1]
#>                 above.last = above.th[no.yrs, 1]
#>                 above = above.first
#>                 for (i in 2:(no.yrs - 1)) {
#>                   above = paste(above, ", ", above.th[i, 1], 
#>                     sep = "")
#>                 }
#>                 above.years = paste0(above, " and ", above.last)
#>                 AboveYears = paste0("The index was above the threshold in ", 
#>                   above.years, ". ")
#>             }
#>             below.th = na.omit(xy[xy$Y > Lsd, ])
#>             if (nrow(below.th) == 0) {
#>                 BelowYears = "The index was never below the threshold. "
#>             }
#>             if (nrow(below.th) == 1) {
#>                 below.last = below.th[1, 1]
#>                 BelowYears = paste0("The index was below the threshold in ", 
#>                   below.last, ". ")
#>             }
#>             if (nrow(below.th) == 2) {
#>                 BelowYears = paste0("The index was below the threshold in ", 
#>                   below.th[1, 1], " and ", below.th[2, 1], ". ")
#>             }
#>             if (nrow(below.th) > 2) {
#>                 no.yrs = nrow(below.th)
#>                 below.first = below.th[1, 1]
#>                 below.last = below.th[no.yrs, 1]
#>                 below = below.first
#>                 for (i in 2:(no.yrs - 1)) {
#>                   below = paste0(below, ", ", below.th[i, 1])
#>                 }
#>                 below.years = paste0(below, " and ", below.last)
#>                 BelowYears = paste0("The index below the threshold in ", 
#>                   below.years, ". ")
#>             }
#>             LastYear = paste0("The index was ", round(Y[length(Y)], 
#>                 2), " in ", X[length(X)], ", the most recent year of the time series. ")
#>             AltText = paste0(OpeningStatement, AboveYears, BelowYears, 
#>                 LastYear)
#>         }
#>         prefix = opts_current$get()$label
#>         if (is.null(prefix)) {
#>             prefix <- ""
#>         }
#>         if (is.na(AltTxt.file.name) == TRUE) {
#>             alt.name = TS
#>         }
#>         else {
#>             alt.name = AltTxt.file.name
#>         }
#>         file.name = paste0(AltTxt, prefix, "_", alt.name, "_alt.txt")
#>         fileConn <- file(file.name)
#>         writeLines(AltText, fileConn)
#>         close(fileConn)
#>     }
#> }
#> <bytecode: 0x7fbfcc2224b0>
#> <environment: namespace:ecoidx>